+++ rio/ChangeLog.md
@@ -5,6 +5,7 @@
 * Re-export the `Control.Monad.State.State` and `Control.Monad.State.StateT` types and related computation functions in `RIO.State`.
 * Re-export the `Control.Monad.Writer.Writer` and `Control.Monad.Writer.WriterT` types and related computation functions in `RIO.Writer`.
 * Re-export `pred`, `succ` in `RIO.Partial`.
+* Add `Semigroup` and `Monoid` instances for `RIO`
 
 ## 0.1.8.0